# server/openapi.yaml
openapi: "3.1.0"
info:
  title: BookFlow API
  version: "1.0.0"
  description: |
    BookFlow is an interactive book authoring and reading platform.

    **Base URL:** `https://books.silverflow.ca/api`

    ## Authentication

    Two methods are supported:

    ### 1. Bearer JWT (session token)
    1. Call `POST /auth/login` with `{ email, password }`
    2. Extract `session.access_token` from the response
    3. Send `Authorization: Bearer <token>` on every subsequent request
    Tokens expire after 1 hour; call login again to refresh.

    ### 2. API Key (long-lived, recommended for automation)
    Send `X-API-Key: <your-api-key>` header instead of Authorization.
    Keys are stored as SHA-256 hashes in the `api_keys` table.
    Contact the admin to generate a key.

    ## Roles
    - `owner` — book creator, full control
    - `author` / `editor` / `viewer` — collaborator roles set per book
    - `super_admin` — bypasses all book-level permission checks
servers:
  - url: https://books.silverflow.ca/api
    description: Production
  - url: http://localhost:8682/api
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access_token from POST /auth/login → session.access_token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Long-lived API key. Contact admin to generate one.
  schemas:
    Book:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        subtitle: { type: string, nullable: true }
        description: { type: string, nullable: true }
        cover_image_url: { type: string, nullable: true }
        status:
          type: string
          enum: [draft, published, archived]
        visibility:
          type: string
          enum: [private, shared, public]
        author_id: { type: string, format: uuid }
        slug: { type: string, nullable: true }
        share_token: { type: string, nullable: true }
        word_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Chapter:
      type: object
      properties:
        id: { type: string, format: uuid }
        book_id: { type: string, format: uuid }
        title: { type: string }
        content:
          type: object
          description: TipTap ProseMirror JSON document. Example — { "type":"doc","content":[{"type":"paragraph","content":[{"type":"text","text":"Hello world"}]}] }
        content_text: { type: string, nullable: true, description: Plain text version, auto-extracted from content }
        order_index: { type: integer }
        status:
          type: string
          enum: [draft, published, archived]
        slug: { type: string, nullable: true }
        word_count: { type: integer }
        estimated_read_time_minutes: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Profile:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        display_name: { type: string, nullable: true }
        avatar_url: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        system_role: { type: string }
        created_at: { type: string, format: date-time }
...
    Comment:
      type: object
      properties:
        id: { type: string, format: uuid }
        chapter_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        content: { type: string }
        resolved: { type: boolean }
        resolved_by: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
    Club:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        owner_id: { type: string, format: uuid }
        type: { type: string, enum: [book_club, class, group] }
        visibility: { type: string, enum: [public, private] }
        created_at: { type: string, format: date-time }
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/profile:
    put:
      tags: [Auth]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name: { type: string }
                bio: { type: string }
                avatar_url: { type: string }
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }

  # ────────────────────────────────────────────────────────────────────────────
  # BOOKS
  # ────────────────────────────────────────────────────────────────────────────
  /books:
    get:
      tags: [Books]
      summary: List books
      description: Without auth returns only public books. With auth also returns your own books.
      security:
        - {}
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: search
          schema: { type: string }
          description: Search by title
        - in: query
          name: visibility
          schema: { type: string, enum: [public, private, shared] }
        - in: query
          name: status
          schema: { type: string, enum: [draft, published, archived] }
        - in: query
          name: author_id
          schema: { type: string, format: uuid }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Book' }

    post:
      tags: [Books]
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, example: My Recovery Story }
                subtitle: { type: string }
                description: { type: string }
                visibility:
                  type: string
                  enum: [private, shared, public]
                  default: private
                cover_image_url: { type: string, format: uri }
      responses:
        "201":
          description: Book created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Book' }
        "401":
          $ref: '#/components/responses/Unauthorized'
...
  /books/public:
    get:
      tags: [Books]
      summary: Discover public published books
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: Search query
      responses:
        "200":
          description: Public books
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Book' }
...
  /books/{id}:
    get:
      tags: [Books]
      summary: Get a single book with chapters, settings, and collaborators
      security:
        - {}
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Book with chapters
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Book' }
                  - type: object
                    properties:
                      chapters:
                        type: array
                        items: { $ref: '#/components/schemas/Chapter' }
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Books]
      summary: Update book metadata
      description: Requires owner or author role.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                subtitle: { type: string }
                description: { type: string }
                status: { type: string, enum: [draft, published, archived] }
                visibility: { type: string, enum: [private, shared, public] }
                cover_image_url: { type: string }
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Book' }

    delete:
      tags: [Books]
      summary: Archive (soft-delete) a book
      description: Sets status to archived. Requires owner or author role.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Book archived
  /books/{id}/settings:
    put:
      tags: [Books]
      summary: Update book reader settings
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                allow_reader_highlights: { type: boolean }
                allow_reader_notes: { type: boolean }
                enable_polls: { type: boolean }
                enable_tts: { type: boolean }
                track_progress: { type: boolean }
                enable_ratings: { type: boolean }
                enable_chapter_chat: { type: boolean }
      responses:
        "200":
          description: Settings updated
...
  /books/{id}/progress:
    get:
      tags: [Books]
      summary: Get my reading progress for a book
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Reading progress

    put:
      tags: [Books]
      summary: Update my reading progress
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chapter_id: { type: string, format: uuid }
                scroll_position: { type: number }
                percent_complete: { type: number }
      responses:
        "200":
          description: Progress saved

  # ────────────────────────────────────────────────────────────────────────────
  # PUBLISH
  # ────────────────────────────────────────────────────────────────────────────
...
  /public/books/{slug}:
    get:
      tags: [Publishing]
      summary: Get a published public book by slug
      security: []
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string, example: my-book-title }
      responses:
        "200":
          description: Public book with chapters
        "404":
          $ref: '#/components/responses/NotFound'
...
  /books/{bookId}/collaborators:
    get:
      tags: [Collaborators]
      summary: List collaborators on a book
      parameters:
        - in: path
          name: bookId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of collaborators

    post:
      tags: [Collaborators]
      summary: Invite a collaborator
      parameters:
        - in: path
          name: bookId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role: { type: string, enum: [author, editor, viewer] }
      responses:
        "201":
          description: Invite sent
...
  /invites/accept/{token}:
    post:
      tags: [Collaborators]
      summary: Accept a collaborator invite
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Invite accepted

  # ────────────────────────────────────────────────────────────────────────────
  # COMMENTS
  # ────────────────────────────────────────────────────────────────────────────
  /chapters/{chapterId}/comments:
    get:
      tags: [Comments]
      summary: List comments on a chapter
      parameters:
        - in: path
          name: chapterId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Comment' }

    post:
      tags: [Comments]
      summary: Add a comment to a chapter
      parameters:
        - in: path
          name: chapterId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                anchor_text: { type: string, description: Selected text this comment refers to }
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }
...
  /clubs:
    get:
      tags: [Clubs]
      summary: List clubs/classes I belong to
      responses:
        "200":
          description: Clubs list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Club' }

    post:
      tags: [Clubs]
      summary: Create a club or class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                type: { type: string, enum: [book_club, class, group] }
                visibility: { type: string, enum: [public, private] }
      responses:
        "201":
          description: Club created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Club' }
...
  /books/{bookId}/versions:
    get:
      tags: [Versions]
      summary: List version snapshots for a book
      parameters:
        - in: path
          name: bookId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Versions list

    post:
      tags: [Versions]
      summary: Create a manual version snapshot
      parameters:
        - in: path
          name: bookId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                label: { type: string, example: "v1.0 — Final draft" }
      responses:
        "201":
          description: Version snapshot created
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Get my notifications
      parameters:
        - in: query
          name: unread_only
          schema: { type: boolean }
      responses:
        "200":
          description: Notifications list
...